Add dependencies for AGP 7.0 and 7.1
To enable adding of tests for these versions to Android Studio.

This is the lint and aapt2 artifacts that are loaded dynamically and are
required for a successful sync of some test projects.

This change only adds artifacts, rather than replacing older artifacts
as would be done for compilation dependencies. This is to enable some
targeted sync tests to verify project import from each supported version
of the Android Gradle Plugin. (When we decide to increase the minimum
supported version in Android Studio the tests can be updated and the
older prebuilts can then be removed.)

Bug: 216916629
Test: See I25f3127b5f7c2a4a5d67e4e7625042b2ab83b1c6
Change-Id: Ia89e717243b054b58422489d8db49439d8a53e28