Remove Instance.immediateDominator field
The field is part of an intermediate computation and shouldn't be
retained. Various functionalities that are using it can use
higher-level information like `nextInstanceToGcRoot`.

This change also removes `LeakedActivityAnalyzerTask`, as it was
using this `immediateDominator` field, and the functionality is
obsoleted by a different component on the Studio side. If anyone
ever wants to reimplement this component on the perflib side,
they should do it in terms of `nextInstanceToGcRoot` instead.

This change also splits the tests involving `immediateDominator`
out to the lower-level tests on the generic dominator algorithm.

Bug: 219770168
Test: split existing
Change-Id: I58e7034c4b9299bbbf3905087fd2baa01d70dbe8