Poll sys.boot_completed property in device setup task
Before this patch, AvdSnapshotHandler relies on the
stdout message from emulator and it waits for
"boot completed" message to be appeared. However,
this message is not displayed on certain API levels
such as API 26.

This patch introduces an extra check. It runs
"adb shell getprop sys.boot_completed" command
every 5 seconds, periodically.

Fixes: 221500241
Test: manually
Change-Id: Ie0783f687f08d1873faedc988f5976153fac0e9e