Stop using JarInputStream.getManifest()
This method only works when the manifest is the first entry listed in
a JAR file. When this returns null, it causes a NPE in the
JarOutputStream.

When upgrading to Bazel 5, we start using JDK 11 for build
actions instead of the legacy java toolchain. This may have something to
do with the ordering of the manifest file, as the manifest was at the
end of input jars.

The manifest file still gets written since all entries are copied to the
output stream.

Bug: 220902606
Test: fixes builds using Bazel 5.0
Change-Id: I578d6defd1ccdbe39156df0c6e79e2978103e727