191392406: Don't inherit class/package annotations for CheckResult
@CheckReturnValue, unlike @CheckResult, can be applied not just on
methods, but on classes and packages too, implying a "default" for all
non-void methods.  However, in this case we don't want to also make it
apply for all *subtypes*.  So for an annotation specified on a specific
method, we'll apply it to overrides as well, but for "outer"
annotations, we'll only apply them within the inner context, not in
subclasses.

Fixes: 191392406
Test: Unit test included
Change-Id: I819c2d4848bdce3fd7a76421c3d2ac9342959789