Increase max attempts of findSerial to 20
findSerial method is invoked after the device launcher
starts emulator command asynchronously. findSerial
method waits for the AVD to start and attached to
adb which may take some time especially on a slow
machine.

This patch changes the max attempts to 20 from 10,
which means the maximum waiting time changes to
8 minutes from 3 minutes.

Bug: 216744701
Test: GradleManagedAndroidDeviceLauncherTest
Change-Id: I1a98b963893dcd17a5be95d4078b0e83129773bf