[GMD] ensure proper task order with clean
Ensure that when a gradle command is run with both GMD clean and GMD
test tasks, that the tasks runs in a coherent deterministic order.

In this case the order is clean -> setup -> test.

This removes the case where the tasks were run out of order leading to
errors in the test task

Bug: 226170214
Test: manual
Change-Id: I09ef94d3df77bc3d4eb78b4a2b471675a298b651