Use bazelisk and Bazel 5.1.1
https://github.com/bazelbuild/bazel/releases/tag/5.1.1

This release contains an important fix for remote execution:
https://github.com/bazelbuild/bazel/pull/15181

The PATH environment variable needs to be included for linux, or else
`bazel run` commands will fail to find `bash` in order to run
executables. A future change will remove explicit passing of environment
variables.

Bazelisk expects to find the LocalAppData environment variable for where
to store Bazel downloads. This had to be added for Bazelisk to work.

Bug: n/a
Test: n/a
Change-Id: I5ff925eef37c3648ee10e49f90ae2b846266a7a5