Add support for getting all module classes
you could already get all module classes by using the
project_classes_dirs and project_classes_jars but that was
not very convenient especially if you needed to transform all
classes.

Instead, we are now proposing a new API dedicated to scoped
artifacts. A scope artifact can be scoped to either the current
project or the full dependency list. So far only CLASSES can
be scoped.

The API allows for getting/appending/transforming/replacing as
for non scoped artifacts.

When getting the current value of a scoped artifact, a Task
must provide two input fields, one for a list of jars and
the other for a list of directories. This is necessary as
inter-project publication is done via a FileCollection in
Gradle, we can get either of those on queries.

Test: added unit tests, API tests and integration tests
Bug: N/A
Change-Id: I8833da2b5a545ecdf91d440d6304762a5217ec53